Split DDTestModuleImpl into parent process and child process implementations
