+++ dd-java-agent/agent-ci-visibility/src/main/java/datadog/trace/civisibility/events/TestEventsHandlerImpl.java
@@ -6,15 +6,20 @@ import datadog.trace.api.Config;
 import datadog.trace.api.DisableTestTrace;
 import datadog.trace.api.civisibility.events.TestEventsHandler;
 import datadog.trace.api.civisibility.source.SourcePathResolver;
+import datadog.trace.api.config.CiVisibilityConfig;
 import datadog.trace.bootstrap.instrumentation.api.Tags;
 import datadog.trace.civisibility.DDTestImpl;
+import datadog.trace.civisibility.DDTestModuleChild;
 import datadog.trace.civisibility.DDTestModuleImpl;
+import datadog.trace.civisibility.DDTestModuleParent;
 import datadog.trace.civisibility.DDTestSuiteImpl;
 import datadog.trace.civisibility.codeowners.Codeowners;
 import datadog.trace.civisibility.context.EmptyTestContext;
 import datadog.trace.civisibility.decorator.TestDecorator;
 import datadog.trace.civisibility.source.MethodLinesResolver;
+import datadog.trace.util.Strings;
 import java.lang.reflect.Method;
+import java.net.InetSocketAddress;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
@@ -62,37 +67,80 @@ public class TestEventsHandlerImpl implements TestEventsHandler {
 
     // some framework/build system combinations fire "onTestModuleStart" event, some cannot do it,
     // hence creating a module here
-    testModule =
-        new DDTestModuleImpl(
-            null,
-            moduleName,
-            null,
-            config,
-            null,
-            testDecorator,
-            sourcePathResolver,
-            codeowners,
-            methodLinesResolver,
-            null);
+    testModule = createTestModule();
   }
 
   @Override
   public void onTestModuleStart() {
     // needed to support JVMs that run tests for multiple modules, e.g. Maven in non-forking mode
     if (testModule == null) {
-      testModule =
-          new DDTestModuleImpl(
-              null,
-              moduleName,
-              null,
-              config,
-              null,
-              testDecorator,
-              sourcePathResolver,
-              codeowners,
-              methodLinesResolver,
-              null);
+      testModule = createTestModule();
     }
+  }
+
+  private DDTestModuleImpl createTestModule() {
+    // fallbacks to System.getProperty below are needed for cases when
+    // system variables are set after config was initialized
+
+    Long parentProcessSessionId = config.getCiVisibilitySessionId();
+    if (parentProcessSessionId == null) {
+      String systemProp =
+          System.getProperty(
+              Strings.propertyNameToSystemPropertyName(CiVisibilityConfig.CIVISIBILITY_SESSION_ID));
+      if (systemProp != null) {
+        parentProcessSessionId = Long.parseLong(systemProp);
+      }
+    }
+
+    Long parentProcessModuleId = config.getCiVisibilityModuleId();
+    if (parentProcessModuleId == null) {
+      String systemProp =
+          System.getProperty(
+              Strings.propertyNameToSystemPropertyName(CiVisibilityConfig.CIVISIBILITY_MODULE_ID));
+      if (systemProp != null) {
+        parentProcessModuleId = Long.parseLong(systemProp);
+      }
+    }
+
+    if (parentProcessSessionId == null || parentProcessModuleId == null) {
+      // it is likely that parent process (build system) is not instrumented
+      // since session and module IDs are not provided to us
+      return new DDTestModuleParent(
+          null,
+          moduleName,
+          null,
+          config,
+          null,
+          testDecorator,
+          sourcePathResolver,
+          codeowners,
+          methodLinesResolver,
+          null);
+    }
+
+    InetSocketAddress signalServerAddress = null;
+    String host =
+        System.getProperty(
+            Strings.propertyNameToSystemPropertyName(
+                CiVisibilityConfig.CIVISIBILITY_SIGNAL_SERVER_HOST));
+    String port =
+        System.getProperty(
+            Strings.propertyNameToSystemPropertyName(
+                CiVisibilityConfig.CIVISIBILITY_SIGNAL_SERVER_PORT));
+    if (host != null && port != null) {
+      signalServerAddress = new InetSocketAddress(host, Integer.parseInt(port));
+    }
+
+    return new DDTestModuleChild(
+        parentProcessSessionId,
+        parentProcessModuleId,
+        moduleName,
+        config,
+        testDecorator,
+        sourcePathResolver,
+        codeowners,
+        methodLinesResolver,
+        signalServerAddress);
   }
 
   @Override
